--- src/test/java/demo/annotations/Person.java
@@ -1,8 +1,0 @@
-package demo.annotations;
-
-/**
- * Created by Helly on 2017/04/10.
- */
-public interface Person {
-    void hello();
-}
