--- src/test/java/demo/annotations/PersonInfo.java
@@ -1,17 +1,0 @@
-package demo.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Created by Helly on 2017/04/10.
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface PersonInfo {
-    String name();
-    int age();
-    boolean sex();
-}
